Legion CI: set CC and CXX to GASNET_{CC,CXX}
This commit adds logic to extract the GASNET_CC and GASNET_CXX
settings from the appropriate installed makefile fragment, and
use them to set CC and CXX in the environment when building the
Legion runtime and applications.

Care is taken to use a subshell to avoid actually including the
selected makefile fragment into the main gmake.  This is valuable to
avoid any possible unintended influence on the Legion builds.
However, it is also necessary to deal with the fact that the fragment
doesn't yet exist at the intended location when gmake begins to run.
It is, in fact, installed and consumed within the same recipe.

This is not intended to replace eventual use of either makefile
fragments or pkgconfig by Legion itself.  However, until either of
those happens, this is meant to simulate a user setting CC and CXX in
their environment before configuring and building GASNet and Legion.