gni: document renamed configure options
This commit adds belated documentation for the rename of the three
configure options which contained "-gni-".

This change was present in the 1.30.0 release, but mistakenly omitted
from both the ChangeLog and the conduit README.  Thanks to Elliot
Ronaghan for reporting this issue.

(cherry picked from commit a6b7c748394445d248fbc9adbd1d91c6ab2833d5)