Chapel CI: provide compatibility with "plain" sh
This commit follows up on the removal of bashisms from the Makefile
with the addition of logic to correctly chose between Chapel's two
Bourne-compatible "setchplenv" scripts (one of which uses extensions
found in Bash and zsh but not available in general).

One alternative would be to always use the non-bash one, but since the
bash one is what appears most prominently in Chapel's documentation,
it seems best to use it on most of our systems.

In addition to simply identifying with script to use, there are
minor changes to cd to the Chapel root directory before including
it, as this is a requirement for their support of non-extended Bourne
shells.

(cherry picked from commit 5b4b24017bd617a9dfa9ab72efb352d3f5d12411)