Fix race in gasneti_defaultExchange()
This commit fixes a race in gasneti_pshmnet_bootstrapBarrierPoll(),
which is used exclusively in gasneti_defaultExchange().  The problem
is that a process could still be calling gasneti_AMPoll() after
another has left the barrier and begun sending non-AM traffic on the
channel.  The solution is to add a second (non-polling) barrier to
the end of the polling barrier.

This commit resolves the assertion below (seen in testing of
aries-conduit) when a gasneti_AMPoll() call reads non-AM traffic from
the shared memory channel.

*** FATAL ERROR: Assertion failure at gasneti_AMPSHM_service_incoming_msg() at
[...]/gasnet/gasnet_pshm.c:1221:
(category == gasneti_Short) || (category == gasneti_Medium) || (category == gasneti_Long)