fix bug3626: prioritize lldb on Mac OS X
The way in which Mac OS X handles the mach equivalent of ptrace_attach()
requires some form or authentication.   This can either be an
interactive prompt (dialogue box or a prompt in ssh session), or a
*key-signed* debugger.  By default only the Xcode-provided lldb is
key-signed.  So, it is the only debugger that can easily be configured
not to require user interaction.

(cherry picked from commit eb59b587a9cd466b4a0b26757107d9fc45ca0fd1)