Add git hash to builds from tarballs
This commit extends existing logic in the dist-hook of the top-level
Makefile.am to embed the git-describe output in an ident string.  This
ensures that it will appear in the source and all libgasnet*.a
(including tools).

(cherry picked from commit 161b31a5ab1b8cc14e7df1aef320f9bbfd48d8fd)