aries/gemini: fix an uninitialized variable
This commit fixes an uninitialized variable specific to the
intersection of EX and multi-domain.  Specifically, the multi-domain
code has always knowingly skipped initializing certain AM-related
fields in the per-domain "peer_data".  However, code added in EX to
deal with the separation of the aux segment from the client segment
caused the previously-unused "pe" field to become necessary.