gemini/aries: honor non-default process layouts
This commit corrects overly-simplistic logic in the conduit-specific
bootstrap exchange operation to allow for process layouts other than
the default pure-blocked layout.  The fix is as simple as restoring
the full generality of the corresponding logic in ibv-conduit, from
which this code way derived.

This commit resolves bug 3647 and addresses upcxx issue 84.