tools: fix gdb-based backtrace on OpenBSD
On OpenBSD the kernel prohibits ptrace attach to an ancestor, and this
prevents a system("gdb ...") from the GASNet process or its child.
This commit works-around this problem by introducing (on OpenBSD only)
a second fork() with exit by the child, breaking the chain of ancestry
between the gdb process and its target.

This commit resolves bug 3636.