configure: add a basic smoke test for MPI_CC
Previously if MPI_CC pointed to an valid executable that fails to work,
configure unconditionally died with a somewhat confusing error.

Now we run a basic compile test, and if that fails we print a more
useful warning message and disable MPI support, allowing configure
to continue (unless the user forced MPI with --enable-mpi(-compat)).