mxm: standardize parsing of GASNET_PHYSMEM_MAX
This commit makes mxm-conduit's handling of the GASNET_PHYSMEM_MAX
and GASNET_PHYSMEM_PROBE environment variables more consistent with
that of ibv and aries/gemini.  This includes addition of the configure
options --with-mxm-physmem-max and --{en|dis}able-mxm-physmem-probe
(using logic newly factored to be shared with ibv-conduit) to set
defaults for these environment variables.

The one significant departure from the behavior of other conduits is
that (for legacy compatibility) mxm-conduit will accept the value "0"
as meaning "2/3".

Additionally, mxm-conduit/README is updated to reflect these changes.