Fix GASNETI_SAFE_PROPAGATE behavior for !gasneti_VerboseErrors
Recent warning changes uncovered a latent bug in GASNETI_SAFE_PROPAGATE
dating back to its creation in 2005. This macro is used infrequently by
{smp,udp,mpi}-conduits to propagate encountered errors back to clients.

gasneti_VerboseErrors is an internal setting enabled by default that
controls console output of error propagation information. When enabled,
the GASNETI_SAFE_PROPAGATE macro operated correctly as intended.
However with gasneti_VerboseErrors==0, the macro was failing to return
and thus failed to actually propagate the error as intended!

AFAIK none of our current clients are modifying the default value of
gasneti_VerboseErrors, so this misbehavior was not reachable in
practice. Even with gasneti_VerboseErrors==0, the misbehavior was on
an error path after something else had already gone wrong.

Also take this opportunity to reduce the excessive string buffer overhead of
GASNETI_SAFE_PROPAGATE and remove extraneous newlines to beautify its output a bit.