bug 3666: workaround buggy __has_attribute on Sun C
__has_attribute((__may_alias__)) crashes the Sun C preprocessor and
has the effect of expanding to the wrong answer.

This effectively makes __has_attribute unusable on Sun C, so disable it.

This workaround will only affect use of Sun C as a mismatched compiler,
where we will conservatively disable all attributes. When Sun C is the
configured compiler, we still use configure results to enable the supported
attributes.