Chapel CI: update CHPL_LAUNCHER handling
This commit responds to upstream improvements by making two changes to
the Makefile logic dealing with Chapel launchers:

  + Chapel now has a gasnetrun_ofi.
    So, remove the corresponding case in the chapel-launcher target.
  + Chapel now has correct default launchers for psm and ofi.
    So, remove the corresponding chapel_launcher_[conduit] variables.

We now only nees to special case pami-conduit.

Additionally, the logic in the clone-launcher has been improved to
more clearly identify the problem when attempting to clone to a
launcher that already exists.