fca: fix used-uninitialized issue
This commit protects against a possible gasneti_free(uninit).

I am unable (as is gcc-7) to determine if all code paths through
__fca_comm_new() will initialize "max_info" prior to reaching the free
call.  So, we now initialize to NULL to be safe (noting that this
function runs only at team creation and thus is not worth a more
detailed analysis to avoid the initialization).