fca: suppress a format warning from gcc
This commit expands the size of a temporary buffer to avoid warnings
about sprintf() potentially overrunning the buffer.  We know the range
of the arguments are such that the original length was sufficient, but
devoting 13 additional bytes of stack space is a simple fix.

(cherry picked from commit 328141897609b5b6708bd0766fa03c7196e90124)