Add test{gasnet,cxx}-mismatch to simulate compiler mismatch
This commit adds four new tests to gasnet-tests harness.conf that
have the effect of compiling the gasnet headers under a simulated
compiler mismatch, to help find regressions in that support.

Note this does not test an actual mismatch, as that would require
a second compiler which we don't necessarily have available.