Improved handling of may_alias attribute for gcc-4
This commit modifies the behavior of gasnet_basic.h with respect to
the may_alias attribute (which can impact correctness) when using a
"mismatched" compiler.  The change is to enable use of this attribute
by default for gcc-4.4+, where the previous behavior was an #error.

It is worth nothing that this default applies only when the user has
not explicitly set GASNETT_USE_GCC_ATTRIBUTE_MAYALIAS nor configured
with --enable-conservative-local-copy.  Additionally, for new-enough
gcc versions the __has_attribute() support has already made use of
this attribute the default.

(cherry picked from commit c1cb90c1061d26181cbbe0ce1b1d64c422c8099e)