gemini/aries: lock outside gasnetc_poll_bound_cq
This commit moves lock/unlock of the domain-specific GNI lock out of
gasnetc_poll_bound_cq() and into gasnetc_poll_local_queue(), which is
its only caller.  This is to enable upcoming changes to perform a
portion of the event processing within the same critical section as
the Cq access.