gemini/aries: non-bulk Puts via GNI LOCAL_EVENT
This commit updates the event processing logic to deal with the "dual
events" case in which both a LOCAL and a GLOBAL event are requested
for the same RMDA Put operation, and uses this to revise the approach
take for non-bulk Puts.  Once beyond the size of a bounce buffer,
RDMA w/ local completion is used rather then the previous approach of
blocking for remote completion of all but a final chunk sent with
synchronous local completion (immediate or bounce buffers).
Measurements with testsmall show a meaningful increase in throughput
of non-bulk Puts.