Fix bug 3676: Misidentification of pgc++ 17.10
This commit fixes bug 3676 - "Misidentification of pgc++ 17.10" by
reordering the compiler family probes to check for __clang__ prior to
the signature any non-Clang compiler that may define it.  This order
is the desired one because later tests over-write earlier.

The problem was compilcated by the fact that pgc++ is defining
__clang__, while surprisingly pgcc does not.  I assume the choice of
front-end is related.  However, the reported mismatch is the only
reason I noticed this problem.