tests: work-around bug 3368 on mismatch tests
This commit addresses the three platforms known (at this point in
time) to #error in the mismatch tests due to an inability to build
compiler-based atomics without configure-probed compiler properties.
It does so by using MANUAL_DEFINES to define the same identifiers that
gasnet_basic.h and gasnet_asm.h would otherwise have defined for the
probed compiler(s).

This is a WORK-AROUND to ensure that the compilers can be applied to
the entire code to find additional errors on these platforms.  If not
for that goal, this may have been resolved as a KnownFailure.