Cleanup: remove TCC atomics support (bug 3672)
This commit removes support for native atomics with TCC (Tiny C
Compiler), switching instead to generics.  This is not a
high-performance compiler and it has become the only consumer of one
of the most complex implementation of inline asm for atomics (which
is removed in this commit).

Resolves bug 3672.

(cherry picked from commit 8ad71da8ca87b1c2547545f66fe0fcf70a896960)