fixup: atomics: work-around bug 3674 (in PGI C+...
This commit corrects the work-around for 3674 by moving the version
check to the correct portion of gasnet_asm.h.  Rather than *move* the
logic from the PPC to X86 portion, I have chosen to *copy*.  This is
based on a conservative assumption that the error is CPU independent.

(cherry picked from commit 804e8a355144896c41309ca1f25cf92dc37b7304)