aries/gemini: fix rare Medium payload corruption
This commit fixes an unfortunate typo that could have lead to the
send of an AM Reply corrupting the payload of the Medium request
handler (by reusing it to hold buffer the reply).

This could have occurred for any Reply needing from 129 and 256 bytes
for the combined args + payload (with payload 8-byte aligned).