Chapel CI: reduce disk usage
This commit adds, conditional on a "chapel_clean" harness feature, a
"make clean" in the Chapel build dir.  This can be used on systems
that are particularly tight on space, but is not done by default
because it makes investigation of any failures more difficult.  If
used, this can yield about 0.5G of space savings (dbg + opt combined).