Initial implementation of gex_AD_Create()
This commit completes the initial implementation of gex_AD_Create(),
and is believed to include all of the conduit-independent logic needed
to support a (yet to be written) implementation based on AMs (without
any an off-loaded alternatives).  The work includes basic constructor
and accessor tests in testgasnet.c.

Notably absent are a destructor and any actual atomic operations.