Fix a (so far) harmless typo.
This commit corrects a misspelling in the forward declaration of the
(never defined) struct which underlies the typedef of gex_TM_t.

No compiler I have yet tested has noticed this issue (even lcc, pcc
and nwcc don't complain when given the typedef in isolation).  So, it
is possible that this forward declaration is unnecessary.  However, I
rather take a defensive (paranoid) position to reduce the possibility
of trouble later on.