Detail planned changes to eop/iop interface
This commit adds comments containing "TODO-EX: EOP_INTERFACE" to
describe the scope of work (expected to be) required to make the
conduit-independent EOP/IOP interface sufficiently general for the
remote atomics and new VIS (with their new event categories).

Additionally, these changes may permit the AM-ref implementation of
Put/Get to make proper use of the conduit-independent EOP/IOP
interface.  It currently uses APIs that we've not established as
"fixed", but which appear in the "common" portions of the extended API
implementation of all conduits which use the AM-ref extended.

One incorrect comment on an #endif is also corrected.