Disable remote atomics in smp-conduit
This commit ensures that smp-conduit will not attempt to build the
reference implementation of remote atomics, and that any client
attempting to use remote atomics with smp-conduit will get (at least)
an informative link error.

Additionally, testgasnet is modified to omit testing of remote atomics
calls for smp-conduit.

This is necessary because smp-conduit does not support the
conduit-independent EOP/IOP interface which will be needed by the
reference implementation.