Separate gasnet_ratomic.h from gasnetex.h
This commit makes gasnet_ratomic.h "stand alone" in the same sense as
gasnet_coll.h and gasnet_vis.h.  The anticipated benefit is that the C
compiler will not need to ingest the non-trivial volume of inline
function definitions that will appear in gasnet_ratomic.h except in
compilation units (client and libgasnet) which have an actual need.

This commit creates a new header (gasnet_ratomic_internal.h) to
contain internal-use declarations and definitions (previously in
gasnet_internal.h) which involve types defined in gasnet_ratomic.h.
As one or more conduit-specific implementations of remote atomics are
implemented, it is likely that more functions or types which are
currently private to gasnet_extended_refratomic.c will appear here.