AM-based implementation of remote atomics
This commit adds a "complete" (meaning all combinations of datatype,
opcode and NB/NBI) implementation of the remote atomics using AMs, for
all conduits except smp-conduit (which lacks the required EOP/IOP
support).

This work includes a short-cut for rank==self to use the tools-based
atomics code directly, without the overheads of the loopback AM.

At this point there is not yet any suite of tests to confirm the
completeness or correctness.

There are numerous additional work items required to complete the
remote atomics.  Among them (in no particular):

  + Documentation in GASNet-EX.spec
  + A flag asserting that rank==self (static)
  + Any actual "selection" in gex_AD_Create()
  + Any conduit-specific implementation
  + Fencing flags