Update segment flag naming: GEX_FLAG_{SELF,PEER}_SEG_*
Change the naming convention for the segment flags from :
  GEX_FLAG_{SRC,DST}_*
to:
  GEX_FLAG_{SELF,PEER}_SEG_{UNKNOWN,SOME,BOUND}

Where the new default corresponds to GASNet-1 behavior:
 GEX_FLAG_SELF_SEG_UNKNOWN, GEX_FLAG_PEER_SEG_BOUND

Update and clarify semantic descriptions.

Add semantics to explicitly define the required validity duration
for the segment disposition assertions. These are specified in
generic terms (as the application to RMA seems self-explanatory) and
explicitly clarified for AM where the details are non-obvious.

All segment flags retain [UNIMPLEMENTED] tag for now.