Add stubs for remote atomics trace/stats
This commit creates extended-ref/gasnet_ratomic_fwd.h which initially
contains only an empty definition of GASNETI_RATOMIC_STATS(), and adds
references to that macro in each gasnet_extended_fwd.h.  This new
files is included by gasnetex.h in the same place as the corresponding
files for the VIS and collective subsystems.