aries: native atomics implementation
This commit adds an implementation of remote atomics using the native
AMO support in the Cray Aries NIC (not Gemini).  In support of this
new implementation there are two new files:
  aries-conduit/gasnet_ratomic_extra.h
  aries-conduit/gasnet_ratomic_fwd.h
These are intentionally placed in the aries-conduit directory rather
than gemini-conduit to ensure they apply only to Aries.

The implementation covers offload of all supported operations on the
integer types ({I,U}{32,64}) and float (FLT), where "supported"
excludes the (F)MULT operations which we recently elected to defer.

This implementation contains all the code necessary to support double
(DBL).  However, it is disabled because there is a known issue[1] with
double-precision addition in the Aries NIC.  Until we have the ability
to chose an implementation based on requested operations, this means
that the native support for DBL is disabled (the AM-based code is
used for this type instead).

Support for GEX_FLAG_IMMEDIATE is included.

For testing purposes, one can add GASNETC_BUILD_GNIRATOMIC=0 to
MANUAL_DEFINES (for libary *and* client) to entirely disable this new
code (and use the AM-based code for all cases).

Legacy "GASNETC_GNI_FETCHOP" support has been removed.

[1] We learned of this from Howard Prichard through both private
communication and his comments in the GNI provider for libfabric, and
have confirmed that a current GNI implementation returns an error if
one attempts to use either the fetching or non-fetching FPADD.