Update seg flags in current implementation
* Update seg flag spelling in header definition and tests

* Remove comments from gasnet.h about adding seg flags, which are
  obsoleted by the fact they now default to GASNet-1 semantics.

For now the flags are given fully-distinct bit patterns, to uphold
the specified requirement for distinct bit patterns and allow debug
checking to detect invalid combinations of mutually exclusive flags.

We may eventually want internal macro shortcuts for the masks that
query logical properties of the assertions (eg was bound seg asserted,
was any seg asserted).