Implement gasneti_THUNK_{CLIENT,TM,EP,SEGMENT}()
This commit provides a family of macros to obtain the current "one and
only" instances of Client, TM, EP and Segment.  This is only for use in
contexts where a function cannot YET obtain it from the caller(s).

This is a *temporary* hack and is to be removed at the earliest
opportunity.