Bug3704: include of gasnet.h after gasnetex.h
To enable clients to include gasnet.h and gasnetex.h in either order,
this commit makes minor changes to gasnetex.h to create a secondary
multiple-include-protection symbol/region to contain definitions in
the gasnet_ name space that are meant only for GASNet-1 clients and
internal code, but are not to be defined in GASNet-EX clients.
Currently only gasnet_handlerentry_t falls into this category.