tests: avoid duplicates in run-tests targets
This commit makes two changes to the logic that builds the
$tests_to_run and $tests_to_ignore shell variables.

The textually first change is to add spaces around both arguments in
the search for $$file within $(testprograms_err).  This is just a bit
of paranoia to deal with the theoretical case in which $$file is a
substring of a test in $(testprograms_err).

The textually second change, the main reason for this commit, is an
application of the same approach to avoid duplicates in $tests_to_run.
In this case, the spaces before/after are necessary due to the
presence of pairs like 'testteambarrier' and 'testteam' (in that
order).