Add WarningFilter for testlegacycxx
This commit adds a filter crafted to match only the bogus warning seen
from g++-2.96 on our "EX-distcheck-old" tester.  This is "fragile" in
that it depends on "distcheck-old" appearing in the source file's full
path, but is done under the assumption that we rather have false
positives than filter out a new warning.