testratomic: update for better native coverage
This commit modifies the randomized atomics test in testratomic to honor
a mask of supported operations.  On Aries conduit, the code now OMITS
from testing (and AD constructions) those operations not supported
natively by the Aries NIC.

This ensures we get coverage of all the natively supported operations on
Aries, and could be adapted to other conduit-specific subsets as well.
However, this is noted as a short-term solution because even if these
are kept as defaults, it seems desirable to allow command-line control.