Make language level part of "matching compiler"
This commit adds logic in configure and gasnet_basic to extend our
definition of "matching compiler" to include the same language level
(__STDC_VERSION__ or __cplusplus, as appropriate).

This change ensures that configure probes that depend on language
level (selected, for instance, via options in {C,CXX}FLAGS) are not
mis-applied.  This is especially true of the newly introduced support
for C++ attributes.