ibv: simplify compile-time increase of MaxMedium
This commit "mostly decouples" the setting of GASNETC_BUFSZ (which is
reduced by the header size to yield the MaxMedium), from that of
GASNETC_PUTINMOVE_LIMIT_MAX (which is used to limit piggybacking of
Put data on remote firehose move Requests).

As long as one wishes only to *raise* GASNETC_BUFSZ at compile time
above the default value of 4096, no other setting is required.  If one
wants to reduce the value AND have configured using
--with-segment-everything, then it remains necessary to define a
matching GASNETC_PUTINMOVE_LIMIT_MAX value.