RAtomic: add GEX_FLAG_AD_FAVOR_*
This commit adds implementation, tests, and documentation for the
GEX_FLAG_AD_FAVOR_* family of hints to influence algorithm selection
in gex_AD_Create().

This includes improved default behavior in aries-conduit to use
AM-based remote atomics for the cases of a singleton team or a run
consisting of a single supernode (unless, of course, the client has
passed GEX_FLAG_AD_FAVOR_REMOTE).