aries: allow RMA for AM-based GEX_OP_{SET,GET}
This commit adds the necessary preprocessor definitions to ensure that
GEX_OP_{SET,GET} can be performed via RMA Put and Get operations when
using the AM-based remote atomics (e.g. when the AD includes one or
more operations not supported for offload to the Aries NIC).

Care has been taken to exclude this optimization for the non-default
case of using generic atomics.