aries/gemini: fix use of uninitialized data
This commit ensures that the 'pe' field of peer_data is always set.

Previously this field was only set for peers outside the supernode.
This was sufficient since the field is used only when initiating
GNI-level communication, which could never occur between processes
within the same supernode.  However, with the introduction of Aries
NIC offload of remote atomics there *is* now a need for GNI-level
intra-supernode communication.  This results in need to access the
'pe' field of *any* peer, for use in gasneti_in_auxsegment().

This issue remained hidden because the memory allocated was often
zeroed, and peer_data[i].pe == 0 was "correct enough" because
gasneti_in_auxsegment() would return 0 as expected.  However, when the
uninitialized memory *did* hold a value other than zero it became
possible to index outside gasneti_seginfo_aux[] (and into unmapped
memory).

The issue was further hidden by the fact that DEBUG builds (but not
NDEBUG) used gasneti_calloc() to purposefully zero-initialize the
peer_data array (which is predominantly pointers).  Since this is at
least the *second* time this has hidden a problem, this commit also
removes this distinction.  Now the peer_data array is always allocated
using gasneti_malloc(), and we rely on GASNet't debug malloc for
initializations meant to poison pointers, if any.