NP-AM: optionally use va_list for args handling
This commit refactors the default (currently only) implementation of
Negotiated-Payload AMs to isolate the unfortunate "spill" of arguments
to a dense array at Commit (to pass to an AM Request or Reply
initiation).

The argument spilling is now isolated in a family of calls:
    gasnetc_AM{Request,Reply}{Medium,Long}V()
which take a va_list instead of individual arguments.

The only implementation of the V-suffixed AM initiation calls is
essentially the same as prior to this commit except that a stack
temporary is used instead of the SD variable.  However, the code
is setup to permit conduit-specific implementations (to appear
in future commits).