aries/gemini: fix gex_Token_Info() bug w/o PSHM
This commit corrects an error in gex_Token_Info() results for a
loopback token and the keys GEX_TI_IS_{REQ,LONG} when PSHM is
disabled.  The results for these two keys were always zero, and in a
debug build this lead to an assertion failure on the first loopback
Request.

Regrettably, this bug was present in the 2017.12.0 release.