ampshm: fix bug in debug checks w/ IMMEDIATE Reply
This commit modifies debug checking used in AMPSHM to account for the
fact that GEX_FLAG_IMMEDIATE may result in a Reply call that does not
actually reply.  Following an IMMEDIATE failure, a subsequent retry
would experience an assertion failure in debug builds, because the
token incorrectly indicated that a Reply has already been sent.

Regrettably, this bug has been present since IMMEDIATE was added (thus
in the 2017.6.0 release).  However, we have not yet had testing of the
problematic case (which requires back-pressure to trigger).