Ratomic: define GEX_FLAG_AD_{ACQ,REL} constants
This commit defines the GEX_FLAG_AD_{ACQ,REL} constants and adds the
basic checks for their proper definitions to testgasnet.

However, these flags are currently both ignored (not providing any
fencing).  So, to avoid confusion in the short-term, the .spec file
has been updated to annotate them as [UNIMPLEMENTED].