RAtomic: add 'fences' arg to gasnete_ratomicfn*()
This commit adds an argument to the gasnete_ratomicfn*() family of
inline functions used to perform remote atomic operations via
GASNet-Tools.  This 'fences' argument is passed to the Tools atomics
to request, for instance, ACQ and REL.

All callers have been updated to simply pass 0 for the new argument,
and therefore this commit does *not* add ACQ/REL support to the Remote
Atomics.  However, it is a prerequisite to doing so.