acinclude: "fix" EXTRACT_EXPR for Clang LTO
This commit changes the C code used by GASNET_TRY_CACHE_EXTRACT_EXPR
to prevent removal of the character array by LTO of (at least) Apple's
Clang 9.0.0.

Note this change is *empirically* determined to prevent the variable
we need from being elided by LTO.  This is not provably "robust" in
the face of all conceivable optimizers.