Ratomic: implement (honor) GEX_FLAG_AD_{ACQ,REL}
This commit adds logic to invoke GASNet-Tools atomics with the proper
fencing flags when gex_AD_Op*() is passed the GEX_FLAG_AD_{ACQ,REL}
flags to perform atomics in shared memory.

This is the only necessary change to fully implement the fencing
flags, and this commit adds comments to explain why nothing additional
is required for the network (AM or Aries) atomics.