NP-AM: improve debug checking in PSHM
This commit improves the multiple-reply checking in a debug build by
performing the check in prepare (rather than commit), as soon as any
"IMMEDIATENESS" has been resolved.  As Dan pointed out in suggesting
this change, we are now be able to detect the case of a call to
gex_AM_Reply*() after a gex_AM_PrepareReply*().