RAtomic: GASNETI_WARN_UNUSED_RESULT annotations
This commit adds GASNETI_WARN_UNUSED_RESULT annotations to the
implementation of gex_AD_OpNB*().  As with gex_RMA_{Put,Get}NB() the
reasoning is that one can never safely discard the returned event.
Similarly, we do *not* annotate the NBI versions because the client
may safely ignore the result unless the pass GEX_FLAG_IMMEDIATE.