Restore unused-result warnings
GCC defaults to enabling -Wunused-result, which warns if a caller of a
function marked with attribute warn_unused_result does not use its return value.

This was inadvertently disabled by the -Wunused macro flag we pass to
disable all the other useless warnings in that family that GASNet code
and headers do not support.

However we want the -Wunused-result warning to remain default-enabled (and even provide
function annotations to support clients finding handle-dropping bugs),
so now explicitly re-enable it after -Wno-unused.