NP-AM: code and macro movement for conduit reuse
This commit moves some functions and macro definitions from
gasnet_am.c to gasnet_am.h, where they can be used by conduit-specific
implementations of NP-AM.

Additionally, the "stubs" for conduit additions to the reference
Prepare and Commit are removed, since it now appears that cloning and
modifying the reference is the best approach.  In particular, there
will not be a penalty for a function call from the reference to the
conduit-specific code (in a different source file).