NP-AM: initial unified loopback FP-AM and NP-AM
This commit provides the initial implementation of Negotiated Payload
for loopback, and implements Fixed Payload AMs in terms of a shared
prepare/commit.

This implementation replaces AMPSHM's special case for loopback,
allowing for the removal of the last portion of loopback support
from gasnet_pshm.c.

The NP path is only reachable via PSHM, not yet from smp/nopshm.