ofi: Enable lazy connections for psm2 provider
In version 1.6 of OFI, the psm2 provider is moving from multiplexing all
communication over a single psm2-level endpoint to providing each
ofi-level endpoint with its own psm2 endpoint. This change has
caused the tear-down time to increase. The issue is that when each remote
endpoint is inserted into the address vector, a connection is made for
all the local endpoints. Setting FI_PSM2_LAZY_CONN causes the
connections to be made when two endpoints first communicate, working
around the problem.