Fix non-transpositional Strided check
The 2017.12.0 release contains a check to enforce non-transpositional
(ie GASNet-1) preconditions on strided metadata.

Due to an accidental omission that check was too weak, failing to fully
enforce the necessary condition for stridelevels > 1.

Also, a defect in the fatal error message formatting for the same
check could produce incorrect/garbled output describing a violation.