AM: stub team arg in gasnetc_nbrhd_RequestGeneric
This commit adds the (eventually necessary) 'tm' argument to the
inline function gasnetc_nbrhd_RequestGeneric() to avoid the need
for all-conduit changes in the future.  Associated with this change,
the local variable "jobrank" is introduced to make it marginally more
clear what is currently happening.

Since the argument has been added to an *inline* function, there
should be no cost for any this change in an NDEBUG build.