NP-AM: no malloc/free for small nbrhd NP Long
This commit updates Prepare{Request,Reply}Long() for both PSHM and
true loopback AM to avoid malloc/free of buffer space for NULL
client_buf and small-enough max_length.  This is similar in spirit to
the "packed long" implementation is some conduits.

In the case of PSHM, there is a minimum allocation granule that would
otherwise waste space after the Long header.

In the case of true loopback, the pool of buffers used for medium
payloads (both NP and FP) can be used for Long payloads as well.

In both cases, there is a size beyond which malloc/free are still
used, but the latency of "small-enough" is improved by this
optimization.  This size is about 4KB for PSHM and is equal to the
MAX_MEDIUM for true-loopback.