aries/gemini: specialized NP Medium
This commit provides a GNI-specific implementation of Negotiated
Payload Medium (both Request and Reply).  This includes:

 + Use of conduit's internal buffer when client_buf==NULL to yield
   zero-copy injection.
 + Request payload sizes larger than gex_AM_MaxRequestMedium(...) when
   the allocator state allows it.
 + Request payload sizes less than max_length (but no less than
   min_length) when that is the best the current allocator state can
   provide.

There is currently no "negotiation" of Reply payloads, because the
default environment settings are such that the pool of bounce buffers
used is already exactly sized to MaxReplyMedium(args==0).  There is
opportunity to negotiate downward (to min_length), but that has not
been implemented.