Change Cray compiler to use standard GASNETE_LOOPING_DIMS
Remove code added in May 2004 by 32f6c024c8d02c1d3809e2d8fa4e7c6113ce1d65 that
forced Strided GASNETE_LOOPING_DIMS to 4 for the Cray T3E compiler, to avoid
bugs with the default of using up to 8 looping dimensions on that compiler.

That machine and variant of the compiler are long dead, and modern Cray C
does not demonstrate the bugs motivating this ancient tweak.