NP-AM: maintainability improvements
This commit responds to Dan's PR comment requesting that the consumers
of the new threadinfo pointer fields have comments naming the fields
they access (since these accesses are via computed offsets, not the
field names).  Additionally, this commit adds assertions at the points
of initialization, so that any movement of the fields will not go
unnoticed.