testthreads: fix UB for single-process case
This commit eliminates degenerate (src==dst) calls to Put, Get and
{Request,Reply}Long which would occur for the case of a single
process.

The fix is to select the peer differently in the single-process case,
such that one communicates with the "next thread" within the same
process instead of the "same thread" in the "next process".  However,
this cannot resolve the case with just 1 thread on 1 process.  So,
this commit also out-laws that case.

(cherry picked from commit 3ae634f646b3daac82c2bef9c8e369ac241d3479)