fixup: testthreads: fix UB for single-process case
This commit repairs a faulty back-port of UB fixes in testthreads.c.

Specifically, I was rushed/sloppy and didn't even test compile the
test and therefore missed the use of the variable "numranks" (only
present in the EX version of the test).  So, this commit corrects
those to uses of gasnet_nodes() or a variable with that value.