NP-AM: fix warnings in ..._loopback_commit_inner()
This commit changes the type of the local variable 'numargs' in
gasnetc_loopback_commit_inner() to be 'int' rather than 'unsigned
int'.  This is consistent with the rest of the AM code and avoids
warnings about ">= 0" comparison, as well as other potential warnings
where it is passed to functions expecting 'int'.