Auto-scaling iteration support for tests
Add auto-scaling iterations to testvisperf.

The iteration argument for testvisperf continues to function as before
for integer arguments. However when given an argument that includes a
decimal point (eg "1.0"), the number specifies a target length in seconds for
each measurement, and the iterations are scaled dynamically to perform
a measurement consuming a duration within a factor of 2 of the target
time.

The new default iterations is 0.1 seconds per measurement (which
results in about a 5-minute run with the other default parameters),
and spends roughly the same amount of time collecting each measurement
reported. 0.1 sec per measurement is probably the minimum required
to obtain reasonably precise measurements for a low-noise system.
Higher values will of course result in a longer test and greated precision.

The new default is in sharp contrast to the old default behavior, where
the several-orders-of-magnitude difference in performance of the
measured operations resulted in the vast majority of run time spent
measuring a few very large, sparse operations (to unnecessarily fine
precision) and a very small fraction of time spent on the contiguous or
small-payload operations (resulting in poor precision and noisy results
for those data points).

make run-tests and upcr-harness still pass the same non-default arguments
to generate a much shorter run (with less accurate results).