tests: fix U.B. in testqueue
This commit fixes overlap between source and destination addresses in
testqueue for the case of loopback, by doubling the size of the
segment allocation.  This additionally ensures that in full-duplex
mode the same memory is not simultaneously the source and destination
of Puts or Gets.

(cherry picked from commit 1c377617b4a88825a0d250d07538250917fc9abf)