Set GASNETE_VIS_NPAM default based on conduit NPAM implementation
VIS should only default to using NPAM on conduits with a "real"
NPAM implementation where we expect it to remove a memcpy from the
injection path. On conduits using the reference NPAM implementation,
using NPAM not only still performs the payload memcpy we were trying
to avoid, but additionally introduces a malloc in the critical path.
VIS's FP AM algorithms avoid that malloc cost by amortizing it over
many injections and sometimes folding it into an existing malloc,
and thus should provide better performance for those conduits.

Of course this is just the default and can be overridden using
MANUAL_DEFINES=-DGASNETE_VIS_NPAM=0/1