AM: error-checking payload queries
This commit adds a level of indirection to the AM Max payload queries
which is intended to be used in DEBUG build to perform checks on the
arguments and the conduit's return value.  In an NDEBUG build the
indirection is flattened away by simply name-shifting.

Note that the only check currently implemented is the 512-byte floor.