NP-AM: improved error reporting
This commit improves the reporting of NP-AM usage errors in which the
client has made back-to-back calls to either Prepare or Commit.
Rather than a cryptic assertion failure the case of a known-but-wrong
magic field now results in a more helpful fatalerror.  The case of a
garbage magic still yields an assertion failure.